Allow running agent in debugging mode for specific node task (#4838)
This change introduces a new knob called VSTSAGENT_DEBUG_TASK which can be set to either task GUID or task name and version (like AzureCLIV2). Once set, the agent will execute node process for any pipeline run which includes this task in debug mode, which will wait for the debugger to be attached.